requirements_macos.txt
# Trellis2 MLX requirements (Apple Silicon)
# NO: flash_attn, spconv, torchsparse (CUDA-only)

# Core
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0,<5
safetensors
huggingface_hub
pillow
numpy
easydict

# Mesh processing
trimesh
xatlas
fast-simplification
pygltflib
plyfile
zstandard

# Image processing
opencv-python-headless
kornia
timm

# Server / demo
fastapi
uvicorn[standard]
pydantic>=2.0
gradio>=6,<7

# MLX (Apple Silicon native)
mlx>=0.31.0

# Metal GPU packages (Apple Silicon — install with: pip install --no-build-isolation <url>)
mtldiffrast @ https://github.com/pedronaugusto/mtldiffrast/archive/main.tar.gz
mtlbvh @ https://github.com/pedronaugusto/mtlbvh/archive/main.tar.gz
cumesh @ https://github.com/pedronaugusto/mtlmesh/archive/main.tar.gz
flex_gemm @ https://github.com/pedronaugusto/mtlgemm/archive/main.tar.gz

# Misc
tqdm
imageio
imageio-ffmpeg
scipy
scikit-image
einops
